Continue to add the source code

+++ app/src/main/java/cn/cz/tetris/utils/SPUtils.java
@@ -27,11 +27,11 @@ public class SPUtils {
         return context.getSharedPreferences(SP_NAME, Context.MODE_PRIVATE).getInt(KEY_LEVEL, GameConstants.LEVEL_NORMAL);
     }
 
-    public static void setSpeed(Context context, long speed) {
-        context.getSharedPreferences(SP_NAME, Context.MODE_PRIVATE).edit().putLong(KEY_SPEED, speed).apply();
+    public static void setSpeed(Context context, int speed) {
+        context.getSharedPreferences(SP_NAME, Context.MODE_PRIVATE).edit().putInt(KEY_SPEED, speed).apply();
     }
 
-    public static long getSpeed(Context context) {
-        return context.getSharedPreferences(SP_NAME, Context.MODE_PRIVATE).getLong(KEY_SPEED, GameConstants.SPEED_1);
+    public static int getSpeed(Context context) {
+        return context.getSharedPreferences(SP_NAME, Context.MODE_PRIVATE).getInt(KEY_SPEED, GameConstants.SPEED_1);
     }
 }
